// src/integrations/supabase/types.ts
export type Event = Tables<'events'> & {
    registration_count: number;
    host_rating: number;
    host_profile: {
        id: string;
        full_name: string;
        avatar_url: string;
    };
};